requirements.txt
mysql-connector-python==9.0.0
requests==2.32.3
openai==0.28.0
fitz==0.0.1.dev2
PyMuPDF==1.24.13
pillow==10.4.0
easyocr==1.7.2
pandas==2.2.2